+++ build.sh
@@ -3,7 +3,7 @@
 set -eu
 
 ci_lib() {
-    xcodebuild -project Screenshare.xcodeproj \
+    /usr/bin/xcodebuild -project Screenshare.xcodeproj \
                -scheme "Screenshare" \
                -destination "platform=macOS" \
                CODE_SIGNING_ALLOWED=NO \